consolidate exome and genome code
